requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
asttokens=2.4.1=pyhd8ed1ab_0
bzip2=1.0.8=h5eee18b_6
ca-certificates=2024.7.4=hbcca054_0
comm=0.2.2=pyhd8ed1ab_0
contourpy=1.2.1=pypi_0
cycler=0.12.1=pypi_0
debugpy=1.6.7=py312h6a678d5_0
decorator=5.1.1=pyhd8ed1ab_0
exceptiongroup=1.2.2=pyhd8ed1ab_0
executing=2.0.1=pyhd8ed1ab_0
expat=2.6.2=h6a678d5_0
filelock=3.13.1=pypi_0
fonttools=4.53.1=pypi_0
fsspec=2024.2.0=pypi_0
imageio=2.34.2=pypi_0
importlib-metadata=8.0.0=pyha770c72_0
importlib_metadata=8.0.0=hd8ed1ab_0
ipykernel=6.29.5=pyh3099207_0
ipython=8.26.0=pyh707e725_0
jedi=0.19.1=pyhd8ed1ab_0
jinja2=3.1.3=pypi_0
jupyter_client=8.6.2=pyhd8ed1ab_0
jupyter_core=5.7.2=py312h06a4308_0
kiwisolver=1.4.5=pypi_0
lazy-loader=0.4=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.4.4=h6a678d5_1
libgcc-ng=14.1.0=h77fa898_0
libgomp=14.1.0=h77fa898_0
libsodium=1.0.18=h36c2ea0_1
libstdcxx-ng=11.2.0=h1234567_1
libuuid=1.41.5=h5eee18b_0
markupsafe=2.1.5=pypi_0
matplotlib=3.9.1=pypi_0
matplotlib-inline=0.1.7=pyhd8ed1ab_0
mpmath=1.3.0=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.6.0=pyhd8ed1ab_0
networkx=3.2.1=pypi_0
numpy=2.0.0=pypi_0
nvidia-cublas-cu11=11.11.3.6=pypi_0
nvidia-cuda-cupti-cu11=11.8.87=pypi_0
nvidia-cuda-nvrtc-cu11=11.8.89=pypi_0
nvidia-cuda-runtime-cu11=11.8.89=pypi_0
nvidia-cudnn-cu11=8.7.0.84=pypi_0
nvidia-cufft-cu11=10.9.0.58=pypi_0
nvidia-curand-cu11=10.3.0.86=pypi_0
nvidia-cusolver-cu11=11.4.1.48=pypi_0
nvidia-cusparse-cu11=11.7.5.86=pypi_0
nvidia-nccl-cu11=2.20.5=pypi_0
nvidia-nvtx-cu11=11.8.86=pypi_0
opencv-python=4.10.0.84=pypi_0
openssl=3.3.1=h4ab18f5_1
packaging=24.1=pyhd8ed1ab_0
pandas=2.2.2=pypi_0
parso=0.8.4=pyhd8ed1ab_0
pexpect=4.9.0=pyhd8ed1ab_0
pickleshare=0.7.5=py_1003
pillow=10.2.0=pypi_0
pip=24.0=py312h06a4308_0
platformdirs=4.2.2=pyhd8ed1ab_0
prompt-toolkit=3.0.47=pyha770c72_0
psutil=5.9.0=py312h5eee18b_0
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
pygments=2.18.0=pyhd8ed1ab_0
pyparsing=3.1.2=pypi_0
python=3.12.4=h5148396_1
python-dateutil=2.9.0=pyhd8ed1ab_0
pytz=2024.1=pypi_0
pyzmq=25.1.2=py312h6a678d5_0
readline=8.2=h5eee18b_0
scikit-image=0.24.0=pypi_0
scipy=1.14.0=pypi_0
setuptools=69.5.1=py312h06a4308_0
six=1.16.0=pyh6c4a22f_0
sqlite=3.45.3=h5eee18b_0
stack_data=0.6.2=pyhd8ed1ab_0
sympy=1.12=pypi_0
tifffile=2024.7.2=pypi_0
tk=8.6.14=h39e8969_0
torch=2.3.1+cu118=pypi_0
torchaudio=2.3.1+cu118=pypi_0
torchvision=0.18.1+cu118=pypi_0
tornado=6.4.1=py312h5eee18b_0
traitlets=5.14.3=pyhd8ed1ab_0
typing-extensions=4.9.0=pypi_0
typing_extensions=4.12.2=pyha770c72_0
tzdata=2024.1=pypi_0
wcwidth=0.2.13=pyhd8ed1ab_0
wheel=0.43.0=py312h06a4308_0
xz=5.4.6=h5eee18b_1
zeromq=4.3.5=h6a678d5_0
zipp=3.19.2=pyhd8ed1ab_0
zlib=1.2.13=h5eee18b_1
